Refactor the use o fetch along the code
Some API calls are redefined in all code. So we refactored this behavior and replaced fetch with axios.
Task definition:
Task 4 (Bonus): You may notice a lot of repeated code (especially when it comes to fetch()). How can you improve this? Try to refactor the code and perhaps also improve the logic to make the app even better!